requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pgvector==0.2.4
sqlalchemy==2.0.23
alembic==1.12.1
langchain==0.0.350
langchain-community==0.0.10
langchain-openai==0.0.2
openai==1.3.7
transformers==4.36.2
torch==2.1.1
sentence-transformers==2.2.2
huggingface-hub==0.19.4
pypdf2==3.0.1
python-docx==1.1.0
pandas==2.1.4
numpy==1.24.4
scikit-learn==1.3.2
pydantic==2.5.0
pydantic-settings==2.1.0
